requirements.txt
python-dotenv==0.20.0
python_telegram_bot==13.13
telegram==0.0.1
